Is now able to change the gamemode as well as the map when !mapvote is used, displaying the gamemode under the map name on each button.
If the file "data/lythos_mapvote/maplist.txt" is present then the gamemode(s) for each map is got from there, and only maps defined in there will be chosen from. If it is not present then maps are selected from available files as usual and "defaultGamemode" in the config file is used for the gamemode.

To Do:
  - Basic idiot-proofing; system for removing whitespaces.
  - Fix an issue where maps aren't being added to RevoteBanList.
  - Add in an Aliases.txt file that changes the text on the buttons to look nicer :)